Реализованы: 1) Валидация 2) централизованная обработка ошибок 3) сбо…

+++ controllers/cards.js
@@ -1,4 +1,6 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
 // добавление новой крточки в БД
 module.exports.createCard = async (req, res, next) => {
@@ -26,17 +28,16 @@ module.exports.getCards = async (req, res, next) => {
 module.exports.deleteCard = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.id);
-    if (card.owner._id.toString() !== req.user._id) {
-      const err = new Error('404 Not Found');
-      err.statusCode = 404;
-      return next(err);
+    if (!card) { // проверка наличия удаляемой карты
+      throw new NotFoundError('404 Not Found');
+    }
+    if (card.owner._id.toString() !== req.user._id) { // проверка владельца карты для удаления
+      throw new ForbiddenError('403 Forbidden');
     }
     const cardDelete = await Card.findOneAndRemove(req.params.id);
     return res.status(200).send({ message: 'card deleted:', data: cardDelete });
-  } catch (e) {
-    const err = new Error('404 Not Found');
-    err.statusCode = 404;
-    next(err);
+  } catch (err) {
+    return next(err);
   }
 };
